Added new comment functionality
